refactor(ui): extract SealsReferences from the duplicated sealing machinery
IsInteractive and RemoteComponent carried byte-identical id plumbing —
the property, the fluent setter, the priority-150 id hook — plus
parallel seal calls and missing-id guards. Both now build on one
SealsReferences concern (id + hook + sealRef + requireId); their
trigger conditions and sealed payloads stay their own.
